requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pg8000
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
firebase-admin
openai
google-generativeai
celery
redis
pytesseract
pdf2image
pillow
python-dotenv
httpx
